[ADD] estate: add first features
Since no repository has been created before to record the progress in the module, this commit covers all the features implemented in the Odoo technical course from the beginning to chapter 11.

The implemented functionalities include: creation of the Property, PropertyOffer, PropertyType and PropertyTag models, as well as their corresponding fields and views of menus, trees, forms and searches. In the same way, constraints have been added to validate the fields entered by the user and give logic to the operation of the module.