[IMP] estate: add features from Chapter 12
The list of properties is added to the property type view.

The state of the property is displayed using the 'statusbar' widget.

All lists are displayed by default in a deterministic order and property types can be ordered manually.

The property form view have conditional display of buttons and fields, and tag colors.

The property and offer list views have color decorations.

Offers and tags are editable directly in the list, and the availability date is hidden by default.

The available properties are filtered by default, and searching on the living area returns results where the area is larger than the given number.

There is a stat button on the property type form view which shows the list of all offers related to properties of the given type when it is clicked on.