[IMP] estate: add features from Chapter 13
It is not be possible to delete a property which is not new or canceled.

When an offer is created, the property state change to 'Offer Received'.

It is not be possible to create an offer with a lower price than an existing offer.

The list of available properties linked to a salesperson is be displayed in their user form view.

The .gitignore file is added.